requirements.txt
# Core ML/Audio
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Speech Models
openai-whisper>=20231117
transformers>=4.35.0
datasets>=2.14.0

# Language Identification
speechbrain>=0.5.15

# TTS / Voice Cloning
TTS>=0.22.0

# NLP
nltk>=3.8.1
epitran>=1.24
phonemizer>=3.2.1
gruut>=2.4.0

# Translation
sentencepiece>=0.1.99
indic-nlp-library>=0.92

# Audio Processing
noisereduce>=3.0.0
webrtcvad>=2.0.10

# Evaluation
jiwer>=3.0.0
pesq>=0.0.4

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Report
pylatex>=1.4.2
